+++ src/core/pushNotifications/PushNotificationService.ts
@@ -1,12 +1,13 @@
 import { Platform, Linking } from 'react-native';
 import * as IntentLauncher from 'expo-intent-launcher';
 
+import { aWeekAgo, isDateBefore, now } from '@covid/utils/datetime';
+
 import { IStorageService } from '../LocalStorageService';
 import { IApiClient } from '../api/ApiClient';
 import { PushToken, IPushTokenRemoteClient } from '../types';
 import { TokenInfoResponse, TokenInfoRequest } from '../user/dto/UserAPIContracts';
-import { now, aWeekAgo, isDateBefore } from '../utils/datetime';
-import { isAndroid } from '../utils/platform';
+import { isAndroid } from '../../utils/platform';
 import Analytics, { events } from '../Analytics';
 
 const KEY_PUSH_TOKEN = 'PUSH_TOKEN';
