+++ src/utils/number.ts
@@ -1,5 +1,3 @@
-import { parse } from 'querystring';
-
 export const cleanIntegerVal = (intString: string) => {
   return parseInt(intString.replace(/ /g, ''), 10);
 };
